ghost button size 분리
